update main, database

+++ Database/database_connection.py
@@ -5,9 +5,11 @@ from cryptography.fernet import Fernet
 class Data:
     def __init__(self):
         super(Data, self).__init__()
+        # self.database = None
         self.create_connection()
 
-    def create_connection(self) -> bool:
+    @staticmethod
+    def create_connection() -> bool:
         """
         Creates a database file and tables in the database
         :return: True, if the database has opened
@@ -24,11 +26,11 @@ class Data:
         query.exec("CREATE TABLE IF NOT EXISTS users"
                    "("
                    "ID INTEGER primary key AUTOINCREMENT,"
-                   "Name VARCHAR(30),"
                    "Surname VARCHAR(30),"
+                   "Name VARCHAR(30),"
                    "Patronymic VARCHAR(30),"
                    "Login VARCHAR(50),"
-                   "Password VARCHAR(50),"
+                   "Password VARCHAR(1000),"
                    "Crypt_key VARCHAR(1000),"
                    "Is_admin BOOLEAN"
                    ")")
@@ -73,55 +75,70 @@ class Data:
 
         return True
 
-    def execute_query_with_params(self, sql_query, query_values=None) -> bool:
+    @staticmethod
+    def execute_query_with_params(sql_query, query_values=None) -> bool:
         """
         :param sql_query: it's string with code on SQL.
         :param query_values: It's list of params for sql_query. They already have the correct sequence.
         :return: True, if the query has been added.
                  False, if the query hasn't been added.
         """
+        db = QtSql.QSqlDatabase.addDatabase('QSQLITE')
+        db.setDatabaseName('database.db')
+        if not db.open():
+            return False
+        print(sql_query)
 
         query = QtSql.QSqlQuery()
         query.prepare(sql_query)
 
-        if query_values is not None:
-            for query_value in query_values:
-                query.addBindValue(query_value)
+        success = query.exec()
 
-        succes = query.exec()
+        return success
 
-        return succes
-
-    def Crypt_password(self, password: str, crypt_key: str) -> str:
+    @staticmethod
+    def Crypt_password(password: str, crypt_key: str) -> str:
         """
         :param password.
         :param crypt_key: the key generated to encrypt the password
         :return: encrypted password
         """
-
+        db = QtSql.QSqlDatabase.addDatabase('QSQLITE')
+        db.setDatabaseName('database.db')
+        if not db.open():
+            return False
         cipher = Fernet(crypt_key.encode('utf-8'))
         crypt_password = cipher.encrypt(password.encode('utf-8'))
 
         return crypt_password.decode('utf-8')
 
-    def Decrypt_password(self, password: str, crypt_key: str) -> str:
+    @staticmethod
+    def Decrypt_password(password: str, crypt_key: str) -> str:
         """
         :param password
         :param crypt_key: the key used to decrypt the password stored in the database
         :return: decrypted password
         """
-
+        db = QtSql.QSqlDatabase.addDatabase('QSQLITE')
+        db.setDatabaseName('database.db')
+        if not db.open():
+            return False
         cipher = Fernet(crypt_key.encode('utf-8'))
         decrypt_password = cipher.decrypt(password.encode('utf-8'))
 
         return decrypt_password.decode('utf-8')
 
-    def get_crypt_key(self, user_id: int) -> str:
+    @staticmethod
+    def get_crypt_key(user_id: int) -> str:
         """
         Returns the value of the encryption key for a specific user
         :param user_id:
         :return: the encryption key (string)
         """
+        db = QtSql.QSqlDatabase.addDatabase('QSQLITE')
+        db.setDatabaseName('database.db')
+        if not db.open():
+            return False
         sql_query = QtSql.QSqlQuery()
         sql_query.prepare("SELECT Crypt_key FROM users WHERE ID=?")
         sql_query.addBindValue(user_id)
@@ -133,7 +150,8 @@ class Data:
         else:
             return ''  # Нужно подумать!
 
-    def register_user(self, surname: str, name: str, patronymic: str, login: str, password: str) -> bool:
+    @staticmethod
+    def register_user(surname: str, name: str, patronymic: str, login: str, password: str) -> bool:
         """
         :param surname:
         :param name:
@@ -143,42 +161,85 @@ class Data:
         :return: True, if the user is registered
                  False, if the user isn't registered
         """
-
+        db = QtSql.QSqlDatabase.addDatabase('QSQLITE')
+        db.setDatabaseName('database.db')
+        if not db.open():
+            return False
         crypt_key = Fernet.generate_key().decode('utf-8')
-        encrypted_password = self.Crypt_password(password, crypt_key)
-        sql_query = "INSERT INTO users (Name, Surname, Patronymic, Login, Password, Crypt_key)" \
-                    " VALUES (?, ?, ?, ?, ?, ?)"
-        return self.execute_query_with_params(sql_query,
-                                              [name, surname, patronymic, login, encrypted_password, crypt_key])
+        encrypted_password = Data.Crypt_password(password, crypt_key)
+        sql_query = f"INSERT INTO users (Surname, Name, Patronymic, Login, Password, Crypt_key, Is_admin)" \
+                    f" VALUES ('{surname}', '{name}', '{patronymic}', '{login}', '{encrypted_password}', '{crypt_key}', 0)"
+        return Data.execute_query_with_params(sql_query=sql_query)
 
-    def update_user_query(self, password: str, id_user: int) -> bool:
+    @staticmethod
+    def update_user_query(password: str, id_user: int) -> bool:
         """
         :param password.
         :param id_user.
         :return: True, if the query is updated
                  False, if the query isn't updated
         """
-        crypt_key = self.get_crypt_key(id_user)
-        encrypted_password = self.Crypt_password(password, crypt_key)
+        db = QtSql.QSqlDatabase.addDatabase('QSQLITE')
+        db.setDatabaseName('database.db')
+        if not db.open():
+            return False
+        crypt_key = Data.get_crypt_key(id_user)
+        encrypted_password = Data.Crypt_password(password, crypt_key)
         sql_query = "UPDATE users SET Password=? WHERE ID=?"
 
-        return self.execute_query_with_params(sql_query, [encrypted_password, id_user])
+        return Data.execute_query_with_params(sql_query, [encrypted_password, id_user])
 
-    def delete_user_query(self, id_user: int) -> bool:
+    @staticmethod
+    def delete_user_query(id_user: int) -> bool:
         """
         :param id_user.
         :return: True, if the query is deleted
                  False, if the query isn't deleted
         """
-
+        db = QtSql.QSqlDatabase.addDatabase('QSQLITE')
+        db.setDatabaseName('database.db')
+        if not db.open():
+            return False
         sql_query = "DELETE FROM users WHERE ID=?"
 
-        return self.execute_query_with_params(sql_query, [id_user])
+        return Data.execute_query_with_params(sql_query, [id_user])
 
-    def create_event(self, id_audience: int, title_event: str, time_start_event: str, time_end_event: str, date: str,
+    @staticmethod
+    def search_user(login: str, password: str) -> dict:
+
+        db = QtSql.QSqlDatabase.addDatabase('QSQLITE')
+        db.setDatabaseName('database.db')
+        if not db.open():
+            return {}
+
+        sql_query = QtSql.QSqlQuery()
+        print(sql_query.exec("SELECT ID, Surname, Name, Patronymic, Login, Password, Crypt_key, Is_admin FROM users"))
+
+        while sql_query.next():
+            print('can')
+            crypt_password = Data.Crypt_password(password, sql_query.value(6))
+            if login == str(sql_query.value(4)) and password == str(sql_query.value(5)):
+                print('нашло совпадение!')
+                return {
+                    'id': sql_query.value(0),
+                    'surname': sql_query.value(1),
+                    'name': sql_query.value(2),
+                    'patronymic': sql_query.value(3),
+                    'login': sql_query.value(4),
+                    'password': Data.Decrypt_password(sql_query.value(5), sql_query.value(6)),
+                    'is_admin': sql_query.value(6)
+                }
+
+        return {'hello': 'world'}
+
+    @staticmethod
+    def create_event(id_audience: int, title_event: str, time_start_event: str, time_end_event: str, date: str,
                      responsible: int, category: str, number_participants: int, max_number_participants: int,
                      category_participants: str, datetime_create_event: str) -> bool:
-
+        db = QtSql.QSqlDatabase.addDatabase('QSQLITE')
+        db.setDatabaseName('database.db')
+        if not db.open():
+            return False
         sql_query = "INSERT INTO events (Id_audience, Title_event, Time_start_event, Time_end_event, Date," \
                     "Responsible, Category," \
                     "Max_number_participants," \
@@ -186,12 +247,17 @@ class Data:
                     "Category_participants," \
                     "Datetime_create_event)" \
                     " VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?)"
-        return self.execute_query_with_params(sql_query,
+        return Data.execute_query_with_params(sql_query,
                                               [id_audience, title_event, time_start_event, time_end_event, date,
                                                responsible, category, number_participants, max_number_participants,
                                                category_participants, datetime_create_event])
 
-    def create_media(self, category: str, path: str, id_event: int) -> bool:
+    @staticmethod
+    def create_media(category: str, path: str, id_event: int) -> bool:
+        db = QtSql.QSqlDatabase.addDatabase('QSQLITE')
+        db.setDatabaseName('database.db')
+        if not db.open():
+            return False
         sql_query = "INSERT INTO media (Category, Path, Id_event)" \
                     " VALUES (?, ?, ?)"
-        return self.execute_query_with_params(sql_query, [category, path, id_event])
+        return Data.execute_query_with_params(sql_query, [category, path, id_event])
